\name{well19937.validate}
\alias{well19937.validate}
\title{Checking the installation of the generator}

\description{
  The generator is run with a fixed seed and its output is compared to
  precomputed values.
}

\usage{
  well19937.validate()
}

\value{
  The function prints the result of the test. There is no output value.
}

\examples{
	well19937.validate()
}

